+++ FVM/TimeIntegration.h
@@ -10,50 +10,50 @@ class TimeIntegration
 {
 public:
 	TimeIntegration(int n_nodes);
-	TimeIntegration(int n_nodes, Eigen::VectorXd m_masses);
-	TimeIntegration(int n_nodes, Eigen::VectorXd m_masses, std::vector<int> constraints, Eigen::MatrixXd rest_pos);
+	TimeIntegration(int n_nodes, Eigen::VectorXf m_masses);
+	TimeIntegration(int n_nodes, Eigen::VectorXf m_masses, std::vector<int> constraints, Eigen::MatrixXf rest_pos);
 	~TimeIntegration();
 
-	void simuExplicit(const Eigen::MatrixXd & pos,
-		const Eigen::MatrixXd & vel,
-		const Eigen::MatrixXd & force,
-		const Eigen::VectorXd & mass);
-	void BackEuler(Eigen::MatrixXd & pos,
-		Eigen::MatrixXd & vel,
-		Eigen::MatrixXd & force,
-		Eigen::MatrixXd & K);
-	void BackEuler(Eigen::MatrixXd & pos,
-		Eigen::MatrixXd & vel,
-		Eigen::MatrixXd & force,
-		Eigen::SparseMatrix<double> & K);
+	void simuExplicit(const Eigen::MatrixXf & pos,
+		const Eigen::MatrixXf & vel,
+		const Eigen::MatrixXf & force,
+		const Eigen::VectorXf & mass);
+	void BackEuler(Eigen::MatrixXf & pos,
+		Eigen::MatrixXf & vel,
+		Eigen::MatrixXf & force,
+		Eigen::MatrixXf & K);
+	void BackEuler(Eigen::MatrixXf & pos,
+		Eigen::MatrixXf & vel,
+		Eigen::MatrixXf & force,
+		Eigen::SparseMatrix<float> & K);
 
-	void BackEuler(Eigen::MatrixXd & pos,
-		Eigen::MatrixXd & restPos,
-		Eigen::MatrixXd & vel,
-		Eigen::MatrixXd & force,
-		Eigen::SparseMatrix<double> & K);
+	void BackEuler(Eigen::MatrixXf & pos,
+		Eigen::MatrixXf & restPos,
+		Eigen::MatrixXf & vel,
+		Eigen::MatrixXf & force,
+		Eigen::SparseMatrix<float> & K);
 
-	void setTimeStep(double t) { m_t = t; }
-	Eigen::MatrixXd  getPositions() { return m_positions; }
-	Eigen::MatrixXd  getVelocities() { return m_velocities; }
-	double & getTimeStep() { return m_t; }
+	void setTimeStep(float t) { m_t = t; }
+	Eigen::MatrixXf  getPositions() { return m_positions; }
+	Eigen::MatrixXf  getVelocities() { return m_velocities; }
+	float & getTimeStep() { return m_t; }
 
 private:
-	void addGroundConstraints(double y, Eigen::MatrixXd & pos, Eigen::MatrixXd & vel);
+	void addGroundConstraints(float y, Eigen::MatrixXf & pos, Eigen::MatrixXf & vel);
 
-	Eigen::PardisoLDLT<Eigen::SparseMatrix<double>> m_pardiso_solver;
+	Eigen::PardisoLDLT<Eigen::SparseMatrix<float>> m_pardiso_solver;
 
-	Eigen::MatrixXd m_positions;  // nodes' positions after one time step
-	Eigen::MatrixXd m_velocities; // nodes' velocities after one time step
-	Eigen::VectorXd m_masses;
+	Eigen::MatrixXf m_positions;  // nodes' positions after one time step
+	Eigen::MatrixXf m_velocities; // nodes' velocities after one time step
+	Eigen::VectorXf m_masses;
 
-	double m_dumpingAlpha;
-	double m_dumpingBelta;
+	float m_dumpingAlpha;
+	float m_dumpingBelta;
 
-	double m_t;                   // time step
+	float m_t;                   // time step
 	int n_nodes;
 
 	std::vector<int> m_constraints;
-	Eigen::MatrixXd	 m_rest;
+	Eigen::MatrixXf	 m_rest;
 
 };
